requirements.txt
Django==1.11.6
django-allauth==0.33.0
django-grappelli==2.10.1
djangorestframework==3.7.0
django-celery==3.2.1